pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "driftwatch"
version = "0.1.0"
description = "CLI tool for LLM output evaluation and drift monitoring"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "typer",
    "rich",
    "pyyaml",
    "openai",
    "anthropic",
    "sentence-transformers",
    "jsonschema",
    "numpy",
    "scipy",
    "pydantic",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[project.scripts]
driftwatch = "driftwatch.cli.main:app"

[tool.setuptools.packages.find]
include = ["driftwatch*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
